+++ Source/Daedalus/Utilities/Graph/DelaunayDatastructures.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "Engine.h"
-#include "DataStructures.h"
+#include <Utilities/DataStructures.h>
 
 #include <array>
 #include <deque>
@@ -14,8 +13,8 @@
  */
 namespace utils {
 	namespace delaunay {
-		typedef Vector2<int64> DelaunayId;                   // Positional ID
-		typedef std::pair<DelaunayId, uint64> GhostId;       // Position & local ID aggregate
+		typedef Vector2<int64_t> DelaunayId;                   // Positional ID
+		typedef std::pair<DelaunayId, uint64_t> GhostId;       // Position & local ID aggregate
 	}
 }
 
@@ -23,10 +22,10 @@ namespace std {
 	template <>
 	struct hash<utils::delaunay::GhostId> {
 		size_t operator()(const utils::delaunay::GhostId & g) const {
-			int64 seed = 0;
-			std::hashCombine(seed, hash<utils::Vector2<int64>>()(g.first));
+			int64_t seed = 0;
+			std::hashCombine(seed, hash<utils::Vector2<int64_t>>()(g.first));
 			std::hashCombine(seed, g.second);
-			return seed;
+			return (unsigned) seed;
 		}
 	};
 }
@@ -41,9 +40,9 @@ namespace utils {
 
 		class Vertex {
 		private:
-			uint64 NumFaces;
-			uint64 Id;
-			uint64 ForeignId;      // Local ID of the vertex in the foreign graph
+			uint64_t NumFaces;
+			uint64_t Id;
+			uint64_t ForeignId;      // Local ID of the vertex in the foreign graph
 			Vector2<> Point;
 			bool bIsForeign;
 			
@@ -52,19 +51,19 @@ namespace utils {
 
 			Vertex(
 				const DelaunayId & gid, const Vector2<> & point,
-				const uint64 id, const uint64 fid, bool isForeign,
-				Face * const inf, const uint64 numFaces
+				const uint64_t id, const uint64_t fid, bool isForeign,
+				Face * const inf, const uint64_t numFaces
 			) : GraphOffset(gid), Point(point), IncidentFace(inf),
 				Id(id), ForeignId(fid), bIsForeign(isForeign), NumFaces(numFaces)
 			{}
 
 		public:
-			Vertex(const DelaunayId & gid, const Vector2<> & point, const uint64 id) :
+			Vertex(const DelaunayId & gid, const Vector2<> & point, const uint64_t id) :
 				Vertex(gid, point, id, 0, false, NULL, 0)
 			{}
 			Vertex(
 				const DelaunayId & gid, const Vector2<> & point,
-				const uint64 id, const uint64 fid
+				const uint64_t id, const uint64_t fid
 			) : Vertex(gid, point, id, fid, true, NULL, 0)
 			{}
 			Vertex(const Vertex & copy) :
@@ -72,15 +71,15 @@ namespace utils {
 					copy.bIsForeign, NULL, copy.NumFaces)
 			{}
 
-			uint64 AddFace(Face * const face);
-			uint64 RemoveFace(Face * const face);
+			uint64_t AddFace(Face * const face);
+			uint64_t RemoveFace(Face * const face);
 
 			inline Face * const GetIncidentFace() const { return IncidentFace; }
 			inline const DelaunayId & ParentGraphOffset() const { return GraphOffset; }
 			inline const Vector2<> & GetPoint() const { return Point; }
-			inline uint64 VertexId() const { return Id; };
-			inline uint64 ForeignVertexId() const { return ForeignId; }
-			inline uint64 FaceCount() const { return NumFaces; }
+			inline uint64_t VertexId() const { return Id; };
+			inline uint64_t ForeignVertexId() const { return ForeignId; }
+			inline uint64_t FaceCount() const { return NumFaces; }
 			inline bool IsForeign() const { return bIsForeign; }
 
 			inline bool operator == (const Vertex & other) const {
@@ -113,23 +112,23 @@ namespace utils {
 		class Face {
 		private:
 			bool bIsDegenerate;
-			uint8 NumVertices;
-			uint64 Id;
+			uint8_t NumVertices;
+			uint64_t Id;
 
 		public:
 			std::array<Vertex *, 3> Vertices;        // Vertices of the triangle provided in CW order
 			std::array<Face *, 3> AdjacentFaces;     // Each adjacent face opposite of the vertex provided
 
 			inline bool IsDegenerate() const { return bIsDegenerate; }
-			inline uint8 VertexCount() const { return NumVertices; }
-			inline uint64 FaceId() const { return Id; }
+			inline uint8_t VertexCount() const { return NumVertices; }
+			inline uint64_t FaceId() const { return Id; }
 
 			/** Creates a degenerate face */
-			Face(Vertex * const v1, Vertex * const v2, const uint64 id) :
+			Face(Vertex * const v1, Vertex * const v2, const uint64_t id) :
 				Face(v1, v2, NULL, id)
 			{}
 
-			Face(Vertex * const v1, Vertex * const v2, Vertex * const v3, const uint64 id) :
+			Face(Vertex * const v1, Vertex * const v2, Vertex * const v3, const uint64_t id) :
 				Vertices({{ v1, v2, v3 }}),
 				AdjacentFaces({{ this, this, v3 == NULL ? NULL : this }}),
 				bIsDegenerate(v3 == NULL),
@@ -139,11 +138,11 @@ namespace utils {
 
 			Face(const Face & copy) : Face(copy.Vertices[0], copy.Vertices[1], copy.Vertices[2], copy.Id) {}
 
-			inline uint8 GetCWVertexIndex(const uint8 current) const {
+			inline uint8_t GetCWVertexIndex(const uint8_t current) const {
 				return (current + 1) % NumVertices;
 			}
 
-			inline int8 GetCWVertexIndex(Vertex * const current) const {
+			inline int8_t GetCWVertexIndex(Vertex * const current) const {
 				auto value = FindVertex(current);
 				return value == -1 ? -1 : GetCWVertexIndex(value);
 			}
@@ -153,12 +152,12 @@ namespace utils {
 				return value == -1 ? NULL : Vertices[value];
 			}
 
-			inline uint8 GetCCWVertexIndex(const uint8 current) const {
+			inline uint8_t GetCCWVertexIndex(const uint8_t current) const {
 				// If degenerate, need to return the opposite vertex index instead
 				return (current + (IsDegenerate() ? 1 : 2)) % NumVertices;
 			}
 
-			inline int8 GetCCWVertexIndex(Vertex * const current) const {
+			inline int8_t GetCCWVertexIndex(Vertex * const current) const {
 				auto value = FindVertex(current);
 				return value == -1 ? -1 : GetCCWVertexIndex(value);
 			}
@@ -168,8 +167,8 @@ namespace utils {
 				return value == -1 ? NULL : Vertices[value];
 			}
 
-			int8 FindVertex(Vertex * const vertex) const;
-			int8 FindFace(Face * const face) const;
+			int8_t FindVertex(Vertex * const vertex) const;
+			int8_t FindFace(Face * const face) const;
 			Face * GetAdjacentFaceCW(Vertex * const sharedVertex);
 			Face * GetAdjacentFaceCCW(Vertex * const sharedVertex);
 
@@ -186,9 +185,9 @@ namespace utils {
 			std::vector<Vertex *> HullVertices;
 			bool bIsCollinear;
 
-			uint64 GetSequence(
+			uint32_t GetSequence(
 				std::deque<Vertex *> & deque,
-				const uint64 start, const uint64 end,
+				const uint32_t start, const uint32_t end,
 				const bool isCw) const;
 		public:
 			ConvexHull() : bIsCollinear(false) {}
@@ -201,8 +200,8 @@ namespace utils {
 				return HullVertices.cend();
 			}
 
-			inline Vertex * operator [] (const uint64 index) { return HullVertices[index]; }
-			inline Vertex * const operator [] (const uint64 index) const {
+			inline Vertex * operator [] (const uint32_t index) { return HullVertices[index]; }
+			inline Vertex * const operator [] (const uint32_t index) const {
 				return HullVertices[index];
 			}
 
@@ -216,23 +215,23 @@ namespace utils {
 				return *this;
 			}
 
-			inline uint64 Size() const { return HullVertices.size(); }
+			inline uint32_t Size() const { return HullVertices.size(); }
 
-			inline uint64 GetSequenceCW(
+			inline uint32_t GetSequenceCW(
 				std::deque<Vertex *> & deque,
-				const uint64 start, const uint64 end
+				const uint32_t start, const uint32_t end
 			) const { return GetSequence(deque, start, end, true); }
 
-			inline uint64 GetSequenceCCW(
+			inline uint32_t GetSequenceCCW(
 				std::deque<Vertex *> & deque,
-				const uint64 start, const uint64 end
+				const uint32_t start, const uint32_t end
 			) const { return GetSequence(deque, start, end, false); }
 
-			inline uint64 NextIndex(const uint64 current) const {
+			inline uint32_t NextIndex(const uint32_t current) const {
 				return (current + 1) % Size();
 			}
 
-			inline uint64 PrevIndex(const uint64 current) const {
+			inline uint32_t PrevIndex(const uint32_t current) const {
 				if (current < 1)
 					return Size() - 1;
 				return current - 1;
@@ -243,21 +242,21 @@ namespace utils {
 			 * if the new point is collinear.
 			 */
 			bool AddVertex(Vertex * const vert);
-			int64 FindVertexById(const uint64 id) const;
-			int64 MinIndex(const std::function<double (Vertex * const)> & valueOf) const;
+			int32_t FindVertexById(const uint32_t id) const;
+			uint32_t MinIndex(const std::function<double (Vertex * const)> & valueOf) const;
 
-			int64 LeftVertexIndex() const;
-			int64 RightVertexIndex() const;
-			int64 TopVertexIndex() const;
-			int64 BottomVertexIndex() const;
+			uint32_t LeftVertexIndex() const;
+			uint32_t RightVertexIndex() const;
+			uint32_t TopVertexIndex() const;
+			uint32_t BottomVertexIndex() const;
 
-			uint64 GetRange(const uint64 start, const uint64 end, const bool isCW) const;
+			uint32_t GetRange(const uint32_t start, const uint32_t end, const bool isCW) const;
 		};
 
 		/**
 		 * Returns 1: CW, 0: Colinear, -1: CCW
 		 */
-		inline int8 IsCWWinding(Vertex * const v1, Vertex * const v2, Vertex * const v3) {
+		inline int8_t IsCWWinding(Vertex * const v1, Vertex * const v2, Vertex * const v3) {
 			return FindWinding(v1->GetPoint(), v2->GetPoint(), v3->GetPoint());
 		}
 	}
@@ -269,11 +268,11 @@ namespace utils {
 
 		// These hashmaps are used for fast lookup of vertices and faces by ID, they should
 		// be kept up-to-date with the vertex and face lists.
-		std::unordered_map<uint64, delaunay::Vertex *> IdVertexMap;
-		std::unordered_map<uint64, delaunay::Face *> IdFaceMap;
+		std::unordered_map<uint64_t, delaunay::Vertex *> IdVertexMap;
+		std::unordered_map<uint64_t, delaunay::Face *> IdFaceMap;
 
-		uint64 CurrentFaceId;
-		uint64 CurrentVertexId;
+		uint64_t CurrentFaceId;
+		uint64_t CurrentVertexId;
 
 		delaunay::DelaunayId Offset;
 		
@@ -283,11 +282,11 @@ namespace utils {
 		 * existing face needs to have which adjacency index updated to point to the new
 		 * face.
 		 */
-		std::pair<delaunay::Face *, int8> AdjustNewFaceAdjacencies(
-			delaunay::Face * const newFace, const uint8 pivotIndex);
-		void AdjustRemovedFaceAdjacencies(delaunay::Face * const face, const uint8 pivotIndex);
-		uint64 GetNextFaceId();
-		uint64 GetNextVertexId();
+		std::pair<delaunay::Face *, int8_t> AdjustNewFaceAdjacencies(
+			delaunay::Face * const newFace, const uint8_t pivotIndex);
+		void AdjustRemovedFaceAdjacencies(delaunay::Face * const face, const uint8_t pivotIndex);
+		uint64_t GetNextFaceId();
+		uint64_t GetNextVertexId();
 
 		/**
 		 * Adds a pre-created vertex. The ID should be provide and be unique. The current
@@ -324,15 +323,15 @@ namespace utils {
 			Faces.clear();
 		}
 
-		inline uint64 VertexCount() const { return Vertices.size(); }
-		inline uint64 FaceCount() const { return Faces.size(); }
+		inline uint32_t VertexCount() const { return Vertices.size(); }
+		inline uint32_t FaceCount() const { return Faces.size(); }
 		inline delaunay::DelaunayId GraphOffset() const { return Offset; }
 
 		const std::vector<delaunay::Vertex const *> GetVertices() const;
 		const std::vector<delaunay::Face const *> GetFaces() const;
 		const std::vector<delaunay::Edge> GetUniqueEdges() const;
 
-		delaunay::Vertex * AddVertex(const Vector2<> & point, const uint64 id);
+		delaunay::Vertex * AddVertex(const Vector2<> & point, const uint64_t id);
 
 		delaunay::Face * FindFace(delaunay::Vertex * const v1, delaunay::Vertex * const v2);
 		delaunay::Face * AddFace(delaunay::Vertex * const v1, delaunay::Vertex * const v2);
@@ -352,10 +351,10 @@ namespace std {
 	template <>
 	struct hash<utils::delaunay::Edge> {
 		size_t operator()(const utils::delaunay::Edge & e) const {
-			int64 seed = 0;
+			int64_t seed = 0;
 			std::hashCombine(seed, e.Start->VertexId());
 			std::hashCombine(seed, e.End->VertexId());
-			return seed;
+			return (unsigned) seed;
 		}
 	};
 }
